@@ -1,5 +1,5 @@
-import '../pages/home_page.dart';
-import '../pages/result_page.dart';
+import 'package:bmi_calculator/pages/home_page.dart';
+import 'package:bmi_calculator/pages/result_page.dart';
 
 const String home = '/';
 const String result = '/result';
